+++ galsim/vonkarman.py
@@ -65,7 +65,10 @@ class VonKarman(GSObject):
     the do_delta=True argument to the VonKarman initializer.
 
     @param lam               Wavelength in nanometers
-    @param r0                Fried parameter in meters.
+    @param r0                Fried parameter at specified wavelength `lam` in meters.  Exactly one
+                             of r0 and r0_500 should be specified.
+    @param r0_500            Fried parameter at 500 nm in meters.  Exactly one of r0 and r0_500
+                             should be specified.
     @param L0                Outer scale in meters.  [default: 25.0]
     @param flux              The flux (in photons/cm^2/s) of the profile. [default: 1]
     @param scale_unit        Units assumed when drawing this profile or evaluating xValue, kValue,
@@ -82,9 +85,10 @@ class VonKarman(GSObject):
     @param gsparams          An optional GSParams argument.  See the docstring for GSParams for
                              details. [default: None]
     """
-    _req_params = { "lam" : float, "r0" : float }
-    _opt_params = { "L0" : float, "flux" : float, "scale_unit" : str, "do_delta" : bool }
-    _single_params = []
+    _req_params = { "lam" : float }
+    _opt_params = { "r0" : float, "r0_500" : float, "L0" : float, "flux" : float,
+                    "scale_unit" : str, "do_delta" : bool }
+    _single_params = [ { "r0" : float, "r0_500" : float } ]
     _takes_rng = False
 
     _has_hard_edges = False
@@ -92,13 +96,24 @@ class VonKarman(GSObject):
     #_is_analytic_x = True  # = not do_delta  defined below.
     _is_analytic_k = True
 
-    def __init__(self, lam, r0, L0=25.0, flux=1, scale_unit=arcsec,
+    def __init__(self, lam, r0=None, r0_500=None, L0=25.0, flux=1, scale_unit=arcsec,
                  do_delta=False, suppress_warning=False, gsparams=None):
 
         # We lose stability if L0 gets too large.  This should be close enough to infinity for
         # all practical purposes though.
         if L0 > 1e10:
             L0 = 1e10
+
+        if r0 is not None and r0_500 is not None:
+            raise GalSimIncompatibleValuesError(
+                "Only one of r0 and r0_500 may be specified",
+                r0=r0, r0_500=r0_500)
+        if r0 is None and r0_500 is None:
+            raise GalSimIncompatibleValuesError(
+                "Either r0 or r0_500 must be specified",
+                r0=r0, r0_500=r0_500)
+        if r0_500 is not None:
+            r0 = r0_500 * (lam/500)**1.2
 
         if isinstance(scale_unit, str):
             self._scale_unit = AngleUnit.from_name(scale_unit)
@@ -152,6 +167,10 @@ class VonKarman(GSObject):
     @property
     def r0(self):
         return self._r0
+
+    @property
+    def r0_500(self):
+        return self._r0*(self._lam/500)**(-1.2)
 
     @property
     def L0(self):